+++ modules/admin/controllers/SkillsController.php
@@ -124,9 +124,12 @@ class SkillsController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->identity->id !== 3) {
 
-        return $this->redirect(['index']);
+            $this->findModel($id)->delete();
+
+            return $this->redirect(['index']);
+        }
     }
 
     /**
